Create single input instead of several
