--- src/pages/SignIn/components/ErrorDiv.tsx
@@ -1,2 +1,0 @@
-export const ErrorDiv = ({children}: {children: string}) =>
-  <div className="text-[11px] text-[#d00]">{children}</div>
